[IMP] split l10n_be_coa_multilang (wip)
