+++ l10n_be_coa_multilang/account.py
@@ -33,17 +33,6 @@ class account_account_template(models.Model):
     _inherit = 'account.account.template'
     name = fields.Char(translate=True)
 
-
-class account_tax_template(models.Model):
-    _inherit = 'account.tax.template'
-    name = fields.Char(translate=True)
-
-
-class account_tax_code_template(models.Model):
-    _inherit = 'account.tax.code.template'
-    name = fields.Char(translate=True)
-
-
 class account_chart_template(models.Model):
     _inherit = 'account.chart.template'
     _order = 'name'
@@ -52,14 +41,3 @@ class account_chart_template(models.Model):
     bank_from_template = fields.Boolean(
         string='Banks/Cash from Template',
         help="Generate Bank/Cash accounts and journals from the Templates.")
-
-
-class account_fiscal_position_template(models.Model):
-    _inherit = 'account.fiscal.position.template'
-    name = fields.Char(translate=True)
-    note = fields.Text(translate=True)
-
-
-class account_journal(models.Model):
-    _inherit = 'account.journal'
-    name = fields.Char(translate=True)
