+++ l10n_be_coa_multilang/wizard/__init__.py
@@ -21,6 +21,3 @@
 ##############################################################################
 
 from . import accounting_report_wizard
-from . import l10n_be_vat_declaration
-from . import l10n_be_partner_vat_listing
-from . import l10n_be_vat_intra
