+++ l10n_be_vat_declaration/__init__.py
@@ -20,11 +20,11 @@
 #
 ##############################################################################
 
-from openerp import models, fields
-
-
-class account_fiscal_position(models.Model):
-    _inherit = 'account.fiscal.position'
-
-    name = fields.Char(translate=True)
-    note = fields.Text(translate=True)
+#from . import wizard_multi_charts_accounts
+#from . import res_config
+#from . import account
+#from . import account_fix
+from . import account_tax_fix
+#from . import account_financial_report
+#from . import partner
+#from . import wizard
