+++ l10n_be_vat_declaration/__openerp__.py
@@ -41,12 +41,12 @@
         'account_menuitem.xml',
         'account_account_type_nov.xml',
         'account_account_template_nov.xml',
-        #'account_tax_code_template_nov.xml',
+        'account_tax_code_template_nov.xml',
         'account_chart_template_nov.xml',
-        #'account_tax_template_nov.xml',
-        #'account_fiscal_position_template_nov.xml',
-        #'account_fiscal_position_tax_template_nov.xml',
-        #'account_fiscal_position_account_template_nov.xml',
+        'account_tax_template_nov.xml',
+        'account_fiscal_position_template_nov.xml',
+        'account_fiscal_position_tax_template_nov.xml',
+        'account_fiscal_position_account_template_nov.xml',
         'l10n_be_sequence.xml',
         'wizard_multi_charts_accounts.xml',
         'account_financial_report.xml',
