+++ l10n_be_vat_declaration/account_tax.py
@@ -20,10 +20,14 @@
 #
 ##############################################################################
 
-from . import wizard_multi_charts_accounts
-from . import res_config
-from . import account
-from . import account_fix
-from . import account_financial_report
-from . import partner
-from . import wizard
+from openerp import models, fields
+
+
+class account_tax_template(models.Model):
+    _inherit = 'account.tax.template'
+    name = fields.Char(translate=True)
+
+
+class account_tax_code_template(models.Model):
+    _inherit = 'account.tax.code.template'
+    name = fields.Char(translate=True)
