+++ l10n_be_vat_declaration/fiscal_position.py
@@ -20,10 +20,17 @@
 #
 ##############################################################################
 
-from . import wizard_multi_charts_accounts
-from . import res_config
-from . import account
-from . import account_fix
-from . import account_financial_report
-from . import partner
-from . import wizard
+from openerp import models, fields
+
+
+class account_fiscal_position_template(models.Model):
+    _inherit = 'account.fiscal.position.template'
+    name = fields.Char(translate=True)
+    note = fields.Text(translate=True)
+
+
+class account_fiscal_position(models.Model):
+    _inherit = 'account.fiscal.position'
+
+    name = fields.Char(translate=True)
+    note = fields.Text(translate=True)
