+++ l10n_be_vat_declaration/static/description/index.html
@@ -1,0 +1,119 @@
+<data>
+
+  <section class="oe_container">
+    <div class="oe_row">
+      <h2 class="oe_slogan">Belgium - Multilingual Chart of Accounts (en/nl/fr)</h2>
+      <h3 class="oe_slogan">Multilanguage alternative for the 'l10n_be' belgian accounting module</h3>
+      <p class="oe_mt32">
+        This module activates the following functionality:
+        <ul>
+          <li>Multilanguage support (en/nl/fr) for Chart of Accounts, Taxes and Tax Codes</li>
+          <li>Multilingual accounting templates</li>
+          <li>Update partner titles for commonly used legal entities</li>
+          <li>Support for the NBB/BNB legal Balance and P&amp;L reportscheme including autoconfiguration of the correct financial report entry when creating/changing a general account</li>
+          <li>
+            The setup wizard
+            <ul>
+              <li>allows to select mono- versus multilingual Chart of Accounts</li>
+              <li>allows to select which languages to install</li>
+              <li>copies the CoA, Tax, Tax Code and Fiscal Position translations from the installation templates</li>
+            </ul>
+          </li>
+          <li>
+            Intervat XML VAT declarations
+            <ul>
+              <li>Periodical VAT Declaration</li>
+              <li>Periodical Intracom Declaration</li>
+              <li>Annual Listing of VAT-Subjected Customers</li>
+            </ul>
+          </li>
+          <li>
+            Standard accounting module enhancements
+          </li>
+          <ul>
+            <li>Allow 'Deferral Method' = 'Balance' on Accounts Payable and Receivable (AP/AR) accounts</li>
+            <li>Add constraint to ensure unique Tax Code per Company</li>
+            <li>Replace Tax Object (account.tax) unique on (name, company_id) constraint by unique on (name, description, company_id)</li>
+            <li>Improved multi-company support via 'active' and 'company_id' fields on 'account.account.type'</li>
+          </ul>
+        </ul>
+      </p>
+    </div>
+  </section>
+
+  <section class="oe_spaced oe_container oe_dark">
+    <div class="oe_row oe_spaced">
+      <h2 class="oe_slogan">Configuration</h2>
+      <div class="oe_span12">
+        <p class="oe_mt32">
+          This module has a different approach compared to the 'l10n_be' module for the population of the Chart of Accounts (CoA).
+          <br/><br/>
+          The 'l10n_be' module comes with a fully populated CoA whereas this module will only create the CoA Classes, Groups
+          and a strict minimum set of general accounts.
+          <br/><br/>
+          In order to have a fully populated CoA, you have to import the customer's
+          CoA after the installation of this module.
+          <br/><br/>
+          As an alternative, you can first install the 'l10n_be' module to get a
+          fully populated CoA.
+          <br/><br/>
+          Afterwards you can uninstall 'l10n_be' and install this module.
+        </p>
+      </div>
+    </div>
+  </section>
+
+  <section class="oe_container">
+    <div class="oe_row oe_spaced">
+      <h2 class="oe_slogan">Credits</h2>
+      <div class="oe_span6">
+        <p class="oe_mt32">
+          Author:
+          <ul>
+            <li>
+              <a href="mailto:info@noviat.com">Noviat</a>
+            </li>
+          </ul>
+        </p>
+      </div>
+      <div class="oe_span6">
+        <p class="oe_mt32">
+          Contributors:
+          <ul>
+            <li>
+              <a href="mailto:info@acsone.eu">acsone</a>
+            </li>
+          </ul>
+        </p>
+      </div>
+    </div>
+  </section>
+
+  <section class="oe_container oe_dark">
+    <div class="oe_row oe_spaced">
+      <div class="oe_span12">
+        <h2 class="oe_slogan">Maintainer</h2>
+      </div>
+      <div class="oe_span8">
+        <p class="oe_mt32">
+          This module is maintained by the OCA.
+        </p>
+        <p>
+          OCA, or the Odoo Community Association, is a nonprofit organization whose mission is to support the collaborative development
+          of Odoo features and promote its widespread use.
+        </p>
+        <p>
+          To contribute to this module, please visit <a href="http://odoo-community.org">http://odoo-community.org.</a>
+        </p>
+      </div>
+      <div class="oe_span4">
+        <a href="http://odoo-community.org">
+          <div class="oe_picture">
+            <img src="oca_logo.png"/>
+          </div>
+        </a>
+      </div>
+    </div>
+  </section>
+
+</data>
