+++ l10n_be_vat_declaration/view/report_l10nbevatdeclaration.xml
