+++ l10n_be_vat_declaration/view/report_l10nbevatintra.xml
