+++ l10n_be_vat_declaration/wizard/l10n_be_partner_vat_listing.py
