+++ l10n_be_vat_declaration/wizard/l10n_be_vat_intra_view.xml
