+++ l10n_be_vat_declaration/wizard/reports.xml
